separate compiler function from aerbackend
