+++ qiskit/providers/aer/backends/aerbackend.py
@@ -22,13 +22,7 @@ import warnings
 from abc import ABC, abstractmethod
 
 from qiskit.circuit import QuantumCircuit, ParameterExpression
-from qiskit.circuit.controlflow import (
-    WhileLoopOp,
-    ForLoopOp,
-    IfElseOp,
-    BreakLoopOp,
-    ContinueLoopOp)
-from qiskit.compiler import assemble, transpile
+from qiskit.compiler import assemble
 from qiskit.providers import BackendV1 as Backend
 from qiskit.providers.models import BackendStatus
 from qiskit.pulse import Schedule
@@ -38,7 +32,7 @@ from qiskit.utils import deprecate_arguments
 from ..aererror import AerError
 from ..jobs import AerJob, AerJobSet, split_qobj
 from ..noise.noise_model import NoiseModel, QuantumErrorLocation
-from ..library.control_flow_instructions import AerMark, AerJump
+from .aer_compiler import compile
 
 # Logger
 logger = logging.getLogger(__name__)
@@ -86,9 +80,6 @@ class AerBackend(Backend, ABC):
         # Set options from backend_options dictionary
         if backend_options is not None:
             self.set_options(**backend_options)
-
-        # Compiler for control flow
-        self._last_flow_id = -1
 
     def _convert_circuit_binds(self, circuit, binds):
         parameterizations = []
@@ -322,217 +313,6 @@ class AerBackend(Backend, ABC):
 
         return Result.from_dict(output)
 
-    def _is_dynamic(self, circuit):
-        if not isinstance(circuit, QuantumCircuit):
-            return False
-        for inst, _, _ in circuit.data:
-            if isinstance(inst, (WhileLoopOp, ForLoopOp, IfElseOp, BreakLoopOp, ContinueLoopOp)):
-                return True
-        return False
-
-    def _decompose_subcircuits(self, circuit, basis_gates=None):
-        if circuit is None:
-            return None
-        if basis_gates is None:
-            basis_gates = self.configuration().basis_gates + ['for_loop', 'while_loop',
-                                                              'if_else', 'break_loop',
-                                                              'continue_loop']
-        circuit = circuit.copy()
-        for inst, _, _ in circuit.data:
-            if isinstance(inst, WhileLoopOp):
-                body, = inst.params
-                inst.params = (self._decompose_subcircuits(body, basis_gates),)
-            elif isinstance(inst, ForLoopOp):
-                loop_parameter, indexset, body = inst.params
-                inst.params = (loop_parameter, indexset,
-                               self._decompose_subcircuits(body, basis_gates))
-            elif isinstance(inst, IfElseOp):
-                true_body, false_body = inst.params
-                inst.params = (self._decompose_subcircuits(true_body, basis_gates),
-                               self._decompose_subcircuits(false_body, basis_gates))
-        circuit = transpile(circuit, basis_gates=basis_gates, optimization_level=0)
-        return circuit
-
-    def _inline_for_loop_op(self, inst):
-        indexset, loop_parameter, body = inst.params
-
-        self._last_flow_id += 1
-        loop_id = self._last_flow_id
-        loop_name = f'loop_{loop_id}'
-
-        ret = QuantumCircuit()
-        for qr in body.qregs:
-            ret.add_register(qr)
-        for cr in body.cregs:
-            ret.add_register(cr)
-        if len(ret.clbits) == 0:
-            ret.add_bits(body.clbits)
-        if len(ret.qubits) == 0:
-            ret.add_bits(body.qubits)
-
-        inlined_body = None
-        break_label = f'{loop_name}_end'
-        for index in indexset:
-            continue_label = f'{loop_name}_{index}'
-            inlined_body = self._inline_circuit(body,
-                                                continue_label,
-                                                break_label)
-            inlined_body = inlined_body.bind_parameters({loop_parameter: index})
-            ret.append(inlined_body,
-                       range(inlined_body.num_qubits),
-                       range(inlined_body.num_clbits))
-            ret.append(AerMark(continue_label, inlined_body.num_qubits),
-                       range(inlined_body.num_qubits), [])
-
-        if inlined_body:
-            ret.append(AerMark(break_label, inlined_body.num_qubits),
-                       range(inlined_body.num_qubits), [])
-        return ret
-
-    def _inline_while_loop_op(self, inst):
-        condition_tuple = inst.condition
-        body, = inst.params
-
-        self._last_flow_id += 1
-        loop_id = self._last_flow_id
-        loop_name = f'while_{loop_id}'
-
-        ret = QuantumCircuit()
-        for qr in body.qregs:
-            ret.add_register(qr)
-        for cr in body.cregs:
-            ret.add_register(cr)
-        if len(ret.clbits) == 0:
-            ret.add_bits(body.clbits)
-        if len(ret.qubits) == 0:
-            ret.add_bits(body.qubits)
-
-        continue_label = f'{loop_name}_continue'
-        loop_start_label = f'{loop_name}_start'
-        break_label = f'{loop_name}_end'
-        inlined_body = self._inline_circuit(body, continue_label, break_label)
-        ret.append(inlined_body,
-                   range(inlined_body.num_qubits),
-                   range(inlined_body.num_clbits))
-        ret.append(AerJump(continue_label, ret.num_qubits),
-                   range(ret.num_qubits), [])
-        ret.append(AerMark(break_label, inlined_body.num_qubits),
-                   range(inlined_body.num_qubits), [])
-        return condition_tuple, continue_label, loop_start_label, break_label, ret
-
-    def _inline_if_else_op(self, inst, continue_label, break_label):
-        condition_tuple = inst.condition
-        true_body, false_body = inst.params
-
-        self._last_flow_id += 1
-        if_id = self._last_flow_id
-        if_name = f'if_{if_id}'
-
-        ret = QuantumCircuit()
-        for qr in true_body.qregs:
-            ret.add_register(qr)
-        for cr in true_body.cregs:
-            ret.add_register(cr)
-        if len(ret.clbits) == 0:
-            ret.add_bits(true_body.clbits)
-        if len(ret.qubits) == 0:
-            ret.add_bits(true_body.qubits)
-
-        if_true_label = f'{if_name}_true'
-        if_else_label = f'{if_name}_else'
-        if_end_label = f'{if_name}_end'
-
-        ret.append(AerMark(if_true_label, ret.num_qubits),
-                   range(ret.num_qubits), [])
-        ret.append(self._inline_circuit(true_body, continue_label, break_label),
-                   range(ret.num_qubits),
-                   range(ret.num_clbits))
-
-        if false_body:
-            ret.append(AerJump(if_end_label, ret.num_qubits),
-                       range(ret.num_qubits), [])
-
-            ret.append(AerMark(if_else_label, ret.num_qubits),
-                       range(ret.num_qubits), [])
-
-            ret.append(self._inline_circuit(false_body, continue_label, break_label),
-                       range(ret.num_qubits),
-                       range(ret.num_clbits))
-
-        ret.append(AerMark(if_end_label, ret.num_qubits),
-                   range(ret.num_qubits), [])
-
-        return condition_tuple, if_true_label, if_else_label if false_body else if_end_label, ret
-
-    def _convert_c_if_args(self, cond_tuple):
-        return [1 if elem is True else 0 if elem is False else elem for elem in cond_tuple]
-
-    def _inline_circuit(self, circ, continue_label, break_label):
-        ret = circ.copy()
-        ret.data = []
-
-        q2i = {}
-        for q in ret.qubits:
-            q2i[q] = len(q2i)
-        c2i = {}
-        for c in ret.clbits:
-            c2i[c] = len(c2i)
-
-        for inst, qargs, cargs in circ.data:
-            if isinstance(inst, ForLoopOp):
-                ret.append(self._inline_for_loop_op(inst),
-                           [q2i[q] for q in qargs],
-                           [c2i[c] for c in cargs])
-            elif isinstance(inst, WhileLoopOp):
-                (cond_tuple, continue_label,
-                 loop_start_label, break_label, body) = self._inline_while_loop_op(inst)
-                c_if_args = self._convert_c_if_args(cond_tuple)
-                ret.append(AerMark(continue_label, ret.num_qubits),
-                           range(ret.num_qubits), [])
-                ret.append(AerJump(loop_start_label, ret.num_qubits).c_if(*c_if_args),
-                           range(ret.num_qubits), [])
-                ret.append(AerJump(break_label, ret.num_qubits),
-                           range(ret.num_qubits), [])
-                ret.append(AerMark(loop_start_label, ret.num_qubits),
-                           range(ret.num_qubits), [])
-                ret.append(body, [q2i[q] for q in qargs], [c2i[c] for c in cargs])
-            elif isinstance(inst, IfElseOp):
-                cond_tuple, true_label, else_label, body = self._inline_if_else_op(inst,
-                                                                                   continue_label,
-                                                                                   break_label)
-                c_if_args = self._convert_c_if_args(cond_tuple)
-                ret.append(AerJump(true_label, ret.num_qubits).c_if(*c_if_args),
-                           range(ret.num_qubits), [])
-                ret.append(AerJump(else_label, ret.num_qubits),
-                           range(ret.num_qubits), [])
-                ret.append(body, [q2i[q] for q in qargs], [c2i[c] for c in cargs])
-            elif isinstance(inst, BreakLoopOp):
-                ret.append(AerJump(break_label, ret.num_qubits),
-                           range(ret.num_qubits), [])
-            elif isinstance(inst, ContinueLoopOp):
-                ret.append(AerJump(continue_label, ret.num_qubits),
-                           range(ret.num_qubits), [])
-            else:
-                ret.append(inst, qargs, cargs)
-
-        return ret
-
-    def _compile(self, circuits):
-        if isinstance(circuits, list):
-            all_static = all([not self._is_dynamic(circuit) for circuit in circuits])
-            if all_static:
-                return circuits
-            basis_gates = self.configuration().basis_gates + ['mark', 'jump']
-            return [transpile(self._inline_circuit(self._decompose_subcircuits(circuit),
-                                                   None, None),
-                              basis_gates=basis_gates) if self._is_dynamic(circuit)
-                    else circuit for circuit in circuits]
-        else:
-            return (transpile(self._inline_circuit(self._decompose_subcircuits(circuits),
-                                                   None, None),
-                              basis_gates=basis_gates) if self._is_dynamic(circuits)
-                    else circuits)
-
     def _assemble(self, circuits, parameter_binds=None, **run_options):
         """Assemble one or more Qobj for running on the simulator"""
         # Remove quantum error instructions from circuits and add to
@@ -549,10 +329,10 @@ class AerBackend(Backend, ABC):
             assemble_binds = []
             assemble_binds.append({param: 1 for bind in parameter_binds for param in bind})
 
-            qobj = assemble(self._compile(circuits), self, parameter_binds=assemble_binds,
+            qobj = assemble(compile(circuits, self), self, parameter_binds=assemble_binds,
                             parameterizations=parameterizations)
         else:
-            qobj = assemble(self._compile(circuits), self)
+            qobj = assemble(compile(circuits, self), self)
 
         # Add options
         for key, val in self.options.__dict__.items():
